Improve the pthread key functions' documentation.
We should probably fix the "destructors are called on null values" bug, but let's at least document the current situation first.

(Why only "probably"? Because code that needs to run on existing versions of Android is going to need to write its destructors to cope with null values, so there's an app compat aspect to this. Technically, it's not impossible that someone has a destructor that _relies_ on getting called even for null too.)

Change-Id: I3a2ead05a91c9e6166cfdd6baf047db3cd2917a8