Clarify clockid_t generation.
For both clock_getcpuclockid() and pthread_getcpuclockid(). These are non-obvious enough that we keep coming back to them, and there's even a stackoverflow question that links to the bionic source as the least unclear existing implementation!

clang optimizes out the unnecessary explicit clearing of CPUCLOCK_PERTHREAD_MASK in the clock_getcpuclockid() -- unnecessary because we just shifted zeroes into the bottom three bits -- so let's leave that in for readability and ease of comparison with pthread_getcpuclockid().

Bug: http://b/376715376
Change-Id: I094c7e270cc88debebd9cd3e9c102614060411a0