Remove #if'ery around __bionic_asm_custom_note_gnu_section().
The intention -- as with the other bionic_asm.h macros -- is that they
should "just work" in the sense of "do nothing" on architectures where
they're not needed.

This requires that we simplify the .bp file, but that seems like a good
thing too. (The transitive include of <features.h> from bionic_asm.h
that caused all this was deliberate to ensure that it's usable from
assembler.)

Change-Id: Ic8cebed1e16d4fc2094985385e5fb315d4ca8e83